remove the important for overlay modal background color + clean some modal components
