+++ packages/slice-machine/lib/builders/SliceBuilder/layout/Header/VariationModal.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from "react";
 import { Formik, Form, Field } from "formik";
 import Modal from "react-modal";
-import Card from "../../../../../components/Card/Default";
+import Card from "@components/Card/Default";
 import Select from "react-select";
 
 import { Text, Box, Button, Label, Input, Flex } from "theme-ui";
-import { AsArray, Variation } from "../../../../models/common/Variation";
+import { AsArray, Variation } from "@models/common/Variation";
 
 const Error = ({ msg }: { msg?: string }) => (
   <Text as="span" sx={{ fontSize: 12, color: "error", mt: "5px", ml: 2 }}>
